Fixes no build do projeto e adicionado o gitignore
