+++ .gitginore
@@ -1,0 +1,4 @@
+*.o
+*.obj
+
+/.vs/
